+++ test_unstructured/partition/test_image.py
@@ -1,10 +1,10 @@
-import pytest
-import requests
 from unittest import mock
 
-import unstructured.partition.pdf as pdf
-import unstructured.partition.image as image
-import unstructured_inference.inference.layout as layout
+import pytest
+import requests
+from unstructured_inference.inference import layout
+
+from unstructured.partition import image, pdf
 
 
 class MockResponse:
@@ -39,7 +39,7 @@ def mock_successful_post(url, **kwargs):
                 "number": 1,
                 "elements": [{"type": "Title", "text": "A Charlie Brown Christmas"}],
             },
-        ]
+        ],
     }
     return MockResponse(status_code=200, response=response)
 
@@ -55,7 +55,7 @@ class MockPageLayout(layout.PageLayout):
                 type="Title",
                 coordinates=[(0, 0), (2, 2)],
                 text="Charlie Brown and the Great Pumpkin",
-            )
+            ),
         ]
 
 
@@ -65,7 +65,7 @@ class MockDocumentLayout(layout.DocumentLayout):
         return [
             MockPageLayout(
                 number=0,
-            )
+            ),
         ]
 
 
@@ -92,13 +92,20 @@ def test_partition_image_api_page_break(monkeypatch, filename="example-docs/exam
     assert partition_image_response[2]["text"] == "A Charlie Brown Christmas"
 
 
-@pytest.mark.parametrize("filename, file", [("example-docs/example.jpg", None), (None, b"0000")])
+@pytest.mark.parametrize(
+    ("filename", "file"),
+    [("example-docs/example.jpg", None), (None, b"0000")],
+)
 def test_partition_image_local(monkeypatch, filename, file):
     monkeypatch.setattr(
-        layout, "process_data_with_model", lambda *args, **kwargs: MockDocumentLayout()
+        layout,
+        "process_data_with_model",
+        lambda *args, **kwargs: MockDocumentLayout(),
     )
     monkeypatch.setattr(
-        layout, "process_file_with_model", lambda *args, **kwargs: MockDocumentLayout()
+        layout,
+        "process_file_with_model",
+        lambda *args, **kwargs: MockDocumentLayout(),
     )
 
     partition_image_response = pdf._partition_pdf_or_image_local(filename, file, is_image=True)
@@ -113,7 +120,8 @@ def test_partition_image_local_raises_with_no_filename():
 
 
 def test_partition_image_api_raises_with_failed_healthcheck(
-    monkeypatch, filename="example-docs/example.jpg"
+    monkeypatch,
+    filename="example-docs/example.jpg",
 ):
     monkeypatch.setattr(requests, "post", mock_successful_post)
     monkeypatch.setattr(requests, "get", mock_unhealthy_get)
@@ -123,7 +131,8 @@ def test_partition_image_api_raises_with_failed_healthcheck(
 
 
 def test_partition_image_api_raises_with_failed_api_call(
-    monkeypatch, filename="example-docs/example.jpg"
+    monkeypatch,
+    filename="example-docs/example.jpg",
 ):
     monkeypatch.setattr(requests, "post", mock_unsuccessful_post)
     monkeypatch.setattr(requests, "get", mock_healthy_get)
@@ -133,11 +142,14 @@ def test_partition_image_api_raises_with_failed_api_call(
 
 
 @pytest.mark.parametrize(
-    "url, api_called, local_called", [("fakeurl", True, False), (None, False, True)]
+    ("url", "api_called", "local_called"),
+    [("fakeurl", True, False), (None, False, True)],
 )
 def test_partition_image(url, api_called, local_called):
     with mock.patch.object(
-        pdf, attribute="_partition_via_api", new=mock.MagicMock()
+        pdf,
+        attribute="_partition_via_api",
+        new=mock.MagicMock(),
     ), mock.patch.object(pdf, "_partition_pdf_or_image_local", mock.MagicMock()):
         image.partition_image(filename="fake.pdf", url=url)
         assert pdf._partition_via_api.called == api_called