+++ unstructured/documents/xml.py
@@ -1,10 +1,9 @@
-from typing import Optional, Union, List
+from typing import List, Optional, Union
 
-import lxml.etree as etree
+from lxml import etree
 
-from unstructured.logger import logger
 from unstructured.documents.base import Document, Page
-
+from unstructured.logger import logger
 
 VALID_PARSERS = Union[etree.HTMLParser, etree.XMLParser, None]
 
@@ -65,7 +64,7 @@ class XMLDocument(Document):
                         "Stylesheets are more commonly parsed with the "
                         "XMLParser. If your HTML does not display properly, try "
                         "`import lxml.etree as etree` and setting "
-                        "`parser=etree.XMLParser()` instead."
+                        "`parser=etree.XMLParser()` instead.",
                     )
                 xslt = etree.parse(self.stylesheet)
                 transform = etree.XSLT(xslt)
@@ -85,6 +84,6 @@ class XMLDocument(Document):
 
     @classmethod
     def from_file(cls, filename, parser: VALID_PARSERS = None, stylesheet: Optional[str] = None):
-        with open(filename, "r+") as f:
+        with open(filename, "r+", encoding="utf8") as f:
             content = f.read()
         return cls.from_string(content, parser=parser, stylesheet=stylesheet)