+++ unstructured/ingest/connector/reddit.py
@@ -1,0 +1,129 @@
+import json
+import os
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+from unstructured.ingest.interfaces import (
+    BaseConnector,
+    BaseConnectorConfig,
+    BaseIngestDoc,
+)
+from unstructured.utils import requires_dependencies
+
+if TYPE_CHECKING:
+    from praw.models import Submission
+
+
+@dataclass
+class SimpleRedditConfig(BaseConnectorConfig):
+    subreddit_name: str
+    client_id: str
+    client_secret: str
+    user_agent: str
+    search_query: str
+    num_posts: int
+
+    # Standard Connector options
+    download_dir: str
+    # where to write structured data
+    output_dir: str
+    preserve_downloads: bool = False
+    re_download: bool = False
+    verbose: bool = False
+
+    def __post_init__(self):
+        if self.num_posts <= 0:
+            raise ValueError("The number of Reddit posts to fetch must be positive.")
+
+
+@dataclass
+class RedditIngestDoc(BaseIngestDoc):
+    config: SimpleRedditConfig = field(repr=False)
+    post: "Submission"
+
+    @property
+    def filename(self) -> Path:
+        return (Path(self.config.download_dir) / f"{self.post.id}.md").resolve()
+
+    def _output_filename(self):
+        return Path(self.config.output_dir) / f"{self.post.id}.json"
+
+    def _create_full_tmp_dir_path(self):
+        self.filename.parent.mkdir(parents=True, exist_ok=True)
+
+    def cleanup_file(self):
+        """Removes the local copy the file (or anything else) after successful processing."""
+        if not self.config.preserve_downloads:
+            if self.config.verbose:
+                print(f"cleaning up {self}")
+            os.unlink(self.filename)
+
+    def get_file(self):
+        """Fetches the "remote" doc and stores it locally on the filesystem."""
+        self._create_full_tmp_dir_path()
+        if not self.config.re_download and self.filename.is_file() and self.filename.stat():
+            if self.config.verbose:
+                print(f"File exists: {self.filename}, skipping download")
+            return
+
+        if self.config.verbose:
+            print(f"fetching {self} - PID: {os.getpid()}")
+        # Write the title plus the body, if any
+        text_to_write = f"# {self.post.title}\n{self.post.selftext}"
+        with open(self.filename, "w", encoding="utf8") as f:
+            f.write(text_to_write)
+
+    def has_output(self):
+        """Determine if structured output for this doc already exists."""
+        output_filename = self._output_filename()
+        return output_filename.is_file() and output_filename.stat()
+
+    def write_result(self):
+        """Write the structured json result for this doc. result must be json serializable."""
+        output_filename = self._output_filename()
+        output_filename.parent.mkdir(parents=True, exist_ok=True)
+        with open(output_filename, "w", encoding="utf8") as output_f:
+            json.dump(self.isd_elems_no_filename, output_f, ensure_ascii=False, indent=2)
+        print(f"Wrote {output_filename}")
+
+
+@requires_dependencies(["praw"], extras="reddit")
+class RedditConnector(BaseConnector):
+    def __init__(self, config: SimpleRedditConfig):
+        from praw import Reddit
+
+        self.config = config
+        self.reddit = Reddit(
+            client_id=config.client_id,
+            client_secret=config.client_secret,
+            user_agent=config.user_agent,
+        )
+        self.cleanup_files = not config.preserve_downloads
+
+    def cleanup(self, cur_dir=None):
+        if not self.cleanup_files:
+            return
+
+        if cur_dir is None:
+            cur_dir = self.config.download_dir
+        sub_dirs = os.listdir(cur_dir)
+        os.chdir(cur_dir)
+        for sub_dir in sub_dirs:
+            # don't traverse symlinks, not that there every should be any
+            if os.path.isdir(sub_dir) and not os.path.islink(sub_dir):
+                self.cleanup(sub_dir)
+        os.chdir("..")
+        if len(os.listdir(cur_dir)) == 0:
+            os.rmdir(cur_dir)
+
+    def initialize(self):
+        pass
+
+    def get_ingest_docs(self):
+        subreddit = self.reddit.subreddit(self.config.subreddit_name)
+        if self.config.search_query:
+            posts = subreddit.search(self.config.search_query, limit=self.config.num_posts)
+        else:
+            posts = subreddit.hot(limit=self.config.num_posts)
+        return [RedditIngestDoc(self.config, post) for post in posts]