+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-## 0.4.16-dev1
+## 0.5.0
+
+### Enhancements
+
+* Add `requires_dependencies` Python decorator to check dependencies are installed before
+instantiating a class or running a function
+
+### Features
+
+* Added Wikipedia connector for ingest cli.
+
+### Fixes
+
+* Fix `process_document` file cleaning on failure
+* Fixes an error introduced in the metadata tracking commit that caused `NarrativeText`
+  and `FigureCaption` elements to be represented as `Text` in HTML documents.
+
+## 0.4.16
 
 ### Enhancements
 
@@ -7,6 +24,14 @@
 ### Features
 
 * Added setup script for Ubuntu
+* Added GitHub connector for ingest cli.
+* Added `partition_md` partitioner.
+* Added Reddit connector for ingest cli.
+
+### Fixes
+
+* Initializes connector properly in ingest.main::MainProcess
+* Restricts version of unstructured-inference to avoid multithreading issue
 
 ## 0.4.15